Automatic Exploratory Data Analysis PDF Report

This function automatically generates a PDF report with different exploratory plots and tables from an MSnSet object.

PomaEDA(
  data,
  imputation = "knn",
  normalization = "log_pareto",
  clean_outliers = TRUE,
  coeff_outliers = 1.5
)

Arguments

data
A MSnSet object. First pData column must be the subject group/type.

imputation
Imputation method. Options are "none", "half_min", "median", "mean", "min" and "knn" (default). If "none", all missing values will be replaced by zero.

normalization
Normalization method. Options are "none", "auto_scaling", "level_scaling", "log_scaling", "log_transformation", "vast_scaling" and "log_pareto" (default).

clean_outliers
Logical. If it's set to TRUE, outliers will be removed from EDA.

coeff_outliers
This value corresponds to the classical 1.5 in \(Q3 + 1.5*IQR\) formula to detect outliers. By changing this value, the permissiveness in outlier detection will change.

Value

An exploratory data analysis PDF report.