Collettion of Imputation Methods for Mass Spectrometry Data

PomaImpute() offers different methods to impute missing values in MS data.

PomaImpute(
  data,
  ZerosAsNA = FALSE,
  RemoveNA = TRUE,
  cutoff = 20,
  method = c("none", "half_min", "median", "mean", "min", "knn")
)

Arguments

data
A MSnSet object. First pData column must be the suject group/type.

ZerosAsNA
Logical that indicates if the zeros in the data are missing values. Default is FALSE.

RemoveNA
Logical that indicates if those metabolites with more than selected cutoff missing values in each group have to be removed. Default is TRUE.

cutoff
Numeric that indicates the percentage of missing values allowed in each group. If one of the groups have less missing values than selected cutoff value, these metabolite will not be removed.

method
Imputation method. Options are c("none", "half_min", "median", "mean", "min", "knn"). If "none", all missing values will be replaced by zero.

Value

A MSnSet object with cleaned data.

References

Armitage, E. G., Godzien, J., Alonso‐Herranz, V., López‐Gonzálvez, Á., & Barbas, C. (2015). Missing value imputation strategies for metabolomics data. Electrophoresis, 36(24), 3050-3060.