Lasso, Ridge and Elasticnet Regularized Generalized Linear Models for Binary Outcomes

PomaLasso() is an implementation of the lasso, ridge and elasticnet regression from glmnet package for binary outcomes.

PomaLasso(
  data,
  alpha = 1,
  ntest = NULL,
  nfolds = 10,
  lambda = NULL,
  labels = FALSE
)

Arguments

data
A MSnSet object. First pData column must be the subject group/type.

alpha
Elasticnet mixing parameter. alpha = 1 is the lasso penalty and alpha = 0 is the ridge penalty. This value must be between 0 and 1.

ntest
Numeric indicating the percentage of observations that will be used as test set. Default is NULL (no test set).

nfolds
Number of folds for CV (default is 10). Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable is nfolds = 3.

lambda
A user supplied lambda sequence. Typical usage is to have the program compute its own lambda sequence based on nlambda and lambda.min.ratio. See ?glmnet::glmnet().

labels
Logical indicating if feature names should be plotted in coefficient plot or not. Default is FALSE.

Value

A list with all results including plots, data frames and the resulting prediction model.

References

Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010). Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software, 33(1), 1-22. URL http://www.jstatsoft.org/v33/i01/.