Implementation of limma R Package on Mass Spectrometry Data

PomaLimma() uses the classical limma package for MS data.

PomaLimma(
  data,
  contrast = NULL,
  covariates = FALSE,
  adjust = "fdr",
  cutoff = NULL
)

Arguments

data
A MSnSet object. First pData column must be the subject group/type.

contrast
A character with the limma comparison. For example, "Group1-Group2" or "control-intervention".

covariates
Logical. If it's set to TRUE all metadata variables stored in pData will be used as covariables. Default = FALSE.

adjust
Multiple comparisons correction method. Options are: "fdr", "holm", "hochberg", "hommel", "bonferroni", "BH" and "BY".

cutoff
Default is NULL. If this value is replaced for a numeric value, the resultant table will contains only those features with an adjusted p-value below selected value.

Value

A data frame with limma results.

References

Matthew E. Ritchie, Belinda Phipson, Di Wu, Yifang Hu, Charity W. Law, Wei Shi, Gordon K. Smyth, limma powers differential expression analyses for RNA-sequencing and microarray studies, Nucleic Acids Research, Volume 43, Issue 7, 20 April 2015, Page e47, https://doi.org/10.1093/nar/gkv007

Examples

data("st000284")

st000284 %>%
  PomaNorm() %>%
  PomaLimma(contrast = "Healthy-CRC", adjust = "fdr")

#> Warning: method argument is empty! log_pareto will be used

#>                                        logFC       AveExpr             t
#> linolenic_acid                  3.338705e-01  3.125000e-05  5.0311801527
#> histidine                       2.641183e-01  2.678571e-05  4.9870408119
#> glyceraldehyde                 -2.927024e-01  2.678571e-05 -4.9840219812
#> hyppuric_acid                  -4.662999e-01  2.090961e-19 -4.3947438009
#> glycochenodeoxycholate         -4.901429e-01  4.603987e-18 -4.0702941619
#> glutamine                       1.539747e-01 -1.479162e-18  3.9234693842
#> malonic_acid_3hba               3.980632e-01  2.331035e-18  3.6900642331
#> x2_deoxyuridine                 1.555290e-01 -2.232143e-05  3.6428761077
#> lysine                          1.741540e-01 -1.785714e-05  3.5743033878
#> methionine                      1.701793e-01 -1.339286e-05  3.5041553734
#> alpha_ketoglutaric_acid         1.378058e-01  1.339286e-05  3.4008083945
#> maleic_acid                    -1.717753e-01 -1.785714e-05 -3.3791951397
#> dimethylglycine                 2.192470e-01 -8.928571e-06  3.1501852075
#> glycocholate                   -4.923497e-01  2.678571e-05 -3.1399748450
#> aspartic_acid                  -2.332522e-01 -8.928571e-06 -3.1065538924
#> hydroxyproline_aminolevulinate -2.232121e-01  2.678571e-05 -3.1024746311
#> glutamic_acid                  -2.255737e-01 -1.339286e-05 -3.0965379447
#> n_acetyl_glycine                2.585804e-01  1.785714e-05  3.0887247392
#> proline                        -1.584859e-01 -4.464286e-06 -3.0837750657
#> linoleic_acid                   2.087946e-01  1.339286e-05  3.0783456764
#> cystathionine                  -2.510037e-01  2.232143e-05 -2.9511803295
#> gama_aminobutyrate              1.401391e-01 -1.785714e-05  2.9042865549
#> allantoin                      -2.250908e-01  2.232143e-05 -2.7627320738
#> oxalic_acid                    -1.494673e-01  1.785714e-05 -2.6954866079
#> kynorenate                     -1.594777e-01 -8.928571e-06 -2.5633503722
#> creatinine                      1.343065e-01 -4.464286e-06  2.5481704776
#> urate                           1.098557e-01 -2.232143e-05  2.4569886282
#> pep                             1.255179e-01 -1.339286e-05  2.4185306027
#> homogentisate                   1.144792e-01 -1.785714e-05  2.3801491160
#> fumaric                        -1.199204e-01  3.125000e-05 -2.3779282905
#> adenylosuccinate               -1.861987e-01 -8.928571e-06 -2.3391516779
#> lactate                        -1.262507e-01  2.232143e-05 -2.3322481491
#> imp                             1.695707e-01 -1.785714e-05  2.2533830050
#> glucose                        -1.112195e-01 -1.785714e-05 -2.2198002980
#> epinephrine                    -9.604613e-02  8.928571e-06 -2.1476878812
#> erythrose                      -1.155804e-01 -8.928571e-06 -2.1022143268
#> x2_aminoadipate                 1.578780e-01 -2.678571e-05  2.0936164886
#> arginine                       -9.902604e-02 -8.928571e-06 -2.0690198387
#> pyruvate                       -1.553624e-01  8.928571e-06 -2.0670814146
#> methyl_succinate                9.399033e-02 -1.785714e-05  2.0663417379
#> l_kynurenine                   -1.297269e-01  4.464286e-06 -2.0241433159
#> oh_phenylpyruvate              -1.043237e-01  8.928571e-06 -2.0155862722
#> margaric_acid                   1.019531e-01 -8.928571e-06  2.0042967384
#> g16bp                          -8.562202e-02  1.339286e-05 -1.9493771229
#> uridine                         1.058519e-01  1.785714e-05  1.9325119155
#> d_leucic_acid                  -2.108981e-01  8.928571e-06 -1.8147562667
#> citraconic_acid                 1.230156e-01  8.928571e-06  1.7981022202
#> reduced_glutatione              7.132217e-02 -1.785714e-05  1.7447411308
#> alanine                         8.297619e-02 -1.785714e-05  1.6209126774
#> salicylurate                   -9.857515e-02 -2.678571e-05 -1.5788323355
#> orotate                        -1.200074e-01 -1.785714e-05 -1.5426045373
#> creatine                       -1.150967e-01  1.339286e-05 -1.5151230964
#> aconitate                       1.029554e-01  4.464286e-06  1.4975555713
#> acetylcholine                   8.890104e-02 -7.318365e-19  1.4533815246
#> adenosine                       1.667277e-01  8.928571e-06  1.4222623526
#> asparagine                      6.288765e-02 -4.017857e-05  1.4116500580
#> xanthurenate                    7.492708e-02 -8.928571e-06  1.3746064395
#> x1_methyladenosine             -7.656622e-02  1.785714e-05 -1.3641160143
#> xanthosine                     -1.108504e-01  1.339286e-05 -1.3405142443
#> homovanilate                   -6.647470e-02 -4.464286e-06 -1.2747469030
#> cystine                        -4.898810e-02 -8.928571e-06 -1.1672970353
#> ornithine                      -6.610789e-02  8.928571e-06 -1.1585062405
#> threonine                       5.914360e-02 -4.464286e-06  1.1142933032
#> glucoronate                     9.394940e-02  8.928571e-06  1.1097328330
#> pentothenate                    9.603497e-02 -4.464286e-06  1.0904328601
#> cystamine                      -9.636458e-02 -1.339286e-05 -1.0535871835
#> glycerate                       7.605506e-02  1.339286e-05  1.0289720935
#> trimethylamine_n_oxide         -1.003378e-01 -4.910714e-05 -1.0045776831
#> tyrosine                        5.293899e-02 -1.785714e-05  0.9873990798
#> dtmp                            3.000149e-02 -2.678571e-05  0.9856938700
#> glutaric_acid                  -6.364360e-02  2.678571e-05 -0.9655986559
#> malate                          5.969122e-02  4.017857e-05  0.9601885237
#> hypoxanthine                    5.301637e-02  2.232143e-05  0.8198543790
#> valine                          4.159673e-02  1.293298e-18  0.7751495711
#> glycerol_3_p                    5.154167e-02 -2.160660e-18  0.7742692551
#> x4_pyridoxic_acid              -2.308854e-02 -8.928571e-06 -0.7580843680
#> anthranilate                    6.764063e-02 -8.928571e-06  0.7571766732
#> carnitine                       3.289955e-02  5.357143e-05  0.7243459980
#> x5_hydroxytryptophan           -2.294494e-02 -4.464286e-06 -0.6650218684
#> leucine_iso_leucine             3.298214e-02 -1.785714e-05  0.6520963141
#> citrulline                      4.000595e-02 -1.339286e-05  0.6442506040
#> succinate_methylmalonate       -3.329539e-02  4.464286e-06 -0.6287282817
#> xanthine                        4.164286e-02  2.232143e-05  0.6192849610
#> cytosine                        1.212537e-01  2.678571e-05  0.5624988200
#> n2_n2_dimethylguanosine        -3.844122e-02 -1.785714e-05 -0.5613475099
#> acetoacetate                    7.401711e-02 -1.339286e-05  0.5334205924
#> biotin                          2.752753e-02 -1.339286e-05  0.5315739348
#> propionate                      3.710193e-02  2.678571e-05  0.5221166150
#> f16bp_f26bp                    -2.443006e-02 -2.129683e-19 -0.5120360861
#> adipic_acid                    -4.668378e-02  4.464286e-06 -0.4741722157
#> fructose                       -4.170908e-02 -1.138412e-18 -0.4678345170
#> d_ga3p_dhap                    -1.354167e-02 -8.928571e-06 -0.4538173518
#> sorbitol                        2.453646e-02 -1.785714e-05  0.4500618667
#> cytidine                       -3.359375e-02  4.464286e-06 -0.4244509747
#> x3_nitro_tyrosine              -2.059747e-02 -1.785714e-05 -0.4095748747
#> amp                            -1.322842e-02 -8.928571e-06 -0.3954017854
#> betaine                         1.934077e-02 -8.928571e-06  0.3899575894
#> aminoisobutyrate               -1.708780e-02 -3.125000e-05 -0.3577713031
#> pyridoxal_5_p                  -3.698438e-02 -1.785714e-05 -0.2993867640
#> niacinamide                     2.307366e-02 -1.785714e-05  0.2782456867
#> taurine                         1.446652e-02  4.464286e-06  0.2684127267
#> serine                          1.295610e-02  4.646581e-19  0.2674613432
#> x1_methylhistamine              1.443527e-02 -4.569138e-19  0.2572996542
#> shikimic_acid                   2.721280e-02 -8.928571e-06  0.2274545997
#> guanidinoacetate                1.220015e-02  4.464286e-06  0.2229202094
#> guanosine                       3.252530e-02 -1.152352e-17  0.1955115558
#> choline                        -9.316220e-03 -1.785714e-05 -0.1591978208
#> glycine                         6.449405e-03  5.846947e-19  0.1091377506
#> pyroglutamic_acid              -5.657738e-03 -3.125000e-05 -0.1002343784
#> phenylalanine                   4.779018e-03  8.928571e-06  0.0959918024
#> oxaloacetate                    6.468750e-03 -1.785714e-05  0.0904072075
#> inosine                         5.479911e-03  1.339286e-05  0.0421183269
#> tryptophan                      3.422619e-05  3.125000e-05  0.0005645772
#>                                     P.Value    adj.P.Val            B
#> linolenic_acid                 9.945133e-07 4.669589e-05  5.279056385
#> histidine                      1.222403e-06 4.669589e-05  5.084377725
#> glyceraldehyde                 1.239714e-06 4.669589e-05  5.071112895
#> hyppuric_acid                  1.707055e-05 4.822430e-04  2.607495175
#> glycochenodeoxycholate         6.495834e-05 1.468059e-03  1.361928962
#> glutamine                      1.158767e-04 2.182345e-03  0.825226819
#> malonic_acid_3hba              2.810653e-04 4.537197e-03  0.007504176
#> x2_deoxyuridine                3.344842e-04 4.724589e-03 -0.152447713
#> lysine                         4.293793e-04 5.391095e-03 -0.381629751
#> methionine                     5.522521e-04 6.240449e-03 -0.612060691
#> alpha_ketoglutaric_acid        7.944798e-04 8.063917e-03 -0.944090960
#> maleic_acid                    8.563451e-04 8.063917e-03 -1.012399389
#> dimethylglycine                1.852440e-03 1.321702e-02 -1.711910273
#> glycocholate                   1.915393e-03 1.321702e-02 -1.742056961
#> aspartic_acid                  2.135547e-03 1.321702e-02 -1.840107989
#> hydroxyproline_aminolevulinate 2.163962e-03 1.321702e-02 -1.852010016
#> glutamic_acid                  2.205939e-03 1.321702e-02 -1.869305834
#> n_acetyl_glycine               2.262329e-03 1.321702e-02 -1.892022383
#> proline                        2.298738e-03 1.321702e-02 -1.906386111
#> linoleic_acid                  2.339295e-03 1.321702e-02 -1.922117651
#> cystathionine                  3.499577e-03 1.883106e-02 -2.283277780
#> gama_aminobutyrate             4.046508e-03 2.078434e-02 -2.412909633
#> allantoin                      6.204569e-03 3.048332e-02 -2.792519095
#> oxalic_acid                    7.557776e-03 3.558453e-02 -2.966654154
#> kynorenate                     1.101657e-02 4.979489e-02 -3.297098368
#> creatinine                     1.149333e-02 4.995179e-02 -3.334059402
#> urate                          1.476477e-02 6.179328e-02 -3.551711672
#> pep                            1.637612e-02 6.608934e-02 -3.641260988
#> homogentisate                  1.813740e-02 6.871999e-02 -3.729295889
#> fumaric                        1.824425e-02 6.871999e-02 -3.734348832
#> adenylosuccinate               2.020119e-02 7.263181e-02 -3.821852753
#> lactate                        2.056830e-02 7.263181e-02 -3.837287855
#> imp                            2.519590e-02 8.627688e-02 -4.010530428
#> glucose                        2.742696e-02 9.115432e-02 -4.082574084
#> epinephrine                    3.280400e-02 1.059101e-01 -4.233774744
#> erythrose                      3.664296e-02 1.128233e-01 -4.326658591
#> x2_aminoadipate                3.741057e-02 1.128233e-01 -4.344005859
#> arginine                       3.968304e-02 1.128233e-01 -4.393255387
#> pyruvate                       3.986704e-02 1.128233e-01 -4.397112873
#> methyl_succinate               3.993745e-02 1.128233e-01 -4.398583921
#> l_kynurenine                   4.413360e-02 1.211487e-01 -4.481667643
#> oh_phenylpyruvate              4.502872e-02 1.211487e-01 -4.498314083
#> margaric_acid                  4.623321e-02 1.214966e-01 -4.520172166
#> g16bp                          5.248913e-02 1.348016e-01 -4.624813944
#> uridine                        5.454772e-02 1.369754e-01 -4.656384719
#> d_leucic_acid                  7.088880e-02 1.741399e-01 -4.869413492
#> citraconic_acid                7.349708e-02 1.767057e-01 -4.898493792
#> reduced_glutatione             8.239101e-02 1.939622e-01 -4.989914974
#> alanine                        1.064325e-01 2.454464e-01 -5.191727920
#> salicylurate                   1.157749e-01 2.616513e-01 -5.257009443
#> orotate                        1.243276e-01 2.754709e-01 -5.311866618
#> creatine                       1.311404e-01 2.849781e-01 -5.352648505
#> aconitate                      1.356459e-01 2.892072e-01 -5.378342162
#> acetylcholine                  1.475068e-01 3.086716e-01 -5.441651592
#> adenosine                      1.563305e-01 3.211882e-01 -5.485134528
#> asparagine                     1.594301e-01 3.217071e-01 -5.499751861
#> xanthurenate                   1.706166e-01 3.382400e-01 -5.549932598
#> x1_methyladenosine             1.738897e-01 3.387851e-01 -5.563905046
#> xanthosine                     1.814259e-01 3.474767e-01 -5.594955951
#> homovanilate                   2.037095e-01 3.836530e-01 -5.678665996
#> cystine                        2.443221e-01 4.517845e-01 -5.806497046
#> ornithine                      2.478818e-01 4.517845e-01 -5.816463825
#> threonine                      2.663385e-01 4.737085e-01 -5.865462306
#> glucoronate                    2.682951e-01 4.737085e-01 -5.870409170
#> pentothenate                   2.766852e-01 4.810066e-01 -5.891122219
#> cystamine                      2.931985e-01 5.019914e-01 -5.929667065
#> glycerate                      3.045942e-01 5.137185e-01 -5.954686492
#> trimethylamine_n_oxide         3.161760e-01 5.251898e-01 -5.978903616
#> tyrosine                       3.245043e-01 5.251898e-01 -5.995611895
#> dtmp                           3.253388e-01 5.251898e-01 -5.997254835
#> glutaric_acid                  3.352785e-01 5.304530e-01 -6.016404148
#> malate                         3.379878e-01 5.304530e-01 -6.021492803
#> hypoxanthine                   4.131640e-01 6.395553e-01 -6.143570399
#> valine                         4.390626e-01 6.599988e-01 -6.178442111
#> glycerol_3_p                   4.395818e-01 6.599988e-01 -6.179109280
#> x4_pyridoxic_acid              4.491913e-01 6.599988e-01 -6.191241017
#> anthranilate                   4.497337e-01 6.599988e-01 -6.191913850
#> carnitine                      4.696035e-01 6.803230e-01 -6.215710579
#> x5_hydroxytryptophan           5.067154e-01 7.247955e-01 -6.256047852
#> leucine_iso_leucine            5.150025e-01 7.255257e-01 -6.264381392
#> citrulline                     5.200671e-01 7.255257e-01 -6.269360320
#> succinate_methylmalonate       5.301627e-01 7.302158e-01 -6.279033954
#> xanthine                       5.363532e-01 7.302158e-01 -6.284804098
#> cytosine                       5.743344e-01 7.645680e-01 -6.317666606
#> n2_n2_dimethylguanosine        5.751176e-01 7.645680e-01 -6.318300311
#> acetoacetate                   5.942673e-01 7.731526e-01 -6.333275214
#> biotin                         5.955439e-01 7.731526e-01 -6.334238565
#> propionate                     6.021012e-01 7.731526e-01 -6.339119959
#> f16bp_f26bp                    6.091264e-01 7.733852e-01 -6.344226780
#> adipic_acid                    6.358355e-01 7.935489e-01 -6.362521397
#> fructose                       6.403543e-01 7.935489e-01 -6.365446612
#> d_ga3p_dhap                    6.503963e-01 7.935489e-01 -6.371776819
#> sorbitol                       6.530978e-01 7.935489e-01 -6.373440168
#> cytidine                       6.716412e-01 8.073985e-01 -6.384415639
#> x3_nitro_tyrosine              6.825063e-01 8.118232e-01 -6.390496080
#> amp                            6.929198e-01 8.118941e-01 -6.396087701
#> betaine                        6.969356e-01 8.118941e-01 -6.398183297
#> aminoisobutyrate               7.208488e-01 8.311828e-01 -6.409979664
#> pyridoxal_5_p                  7.649205e-01 8.730910e-01 -6.428788520
#> niacinamide                    7.810785e-01 8.744851e-01 -6.434775712
#> taurine                        7.886268e-01 8.744851e-01 -6.437411206
#> serine                         7.893582e-01 8.744851e-01 -6.437661175
#> x1_methylhistamine             7.971819e-01 8.745782e-01 -6.440275748
#> shikimic_acid                  8.202762e-01 8.865649e-01 -6.447369676
#> guanidinoacetate               8.237992e-01 8.865649e-01 -6.448371064
#> guanosine                      8.451684e-01 9.009814e-01 -6.453994926
#> choline                        8.736554e-01 9.226454e-01 -6.460311621
#> glycine                        9.131902e-01 9.447653e-01 -6.466899444
#> pyroglutamic_acid              9.202472e-01 9.447653e-01 -6.467813685
#> phenylalanine                  9.236123e-01 9.447653e-01 -6.468221983
#> oxaloacetate                   9.280438e-01 9.447653e-01 -6.468732519
#> inosine                        9.664416e-01 9.750705e-01 -6.471871190
#> tryptophan                     9.995500e-01 9.995500e-01 -6.472741094