Rank Product/Rank Sum Analysis for Mass Spectrometry Data

PomaRankProd() performs the Rank Product method to identify differential feature concentration/intensity.

PomaRankProd(
  data,
  logged = TRUE,
  logbase = 2,
  paired = NA,
  cutoff = 0.05,
  method = c("pfp", "pval")
)

Arguments

data
A MSnSet object. First pData column must be the suject group/type.

logged
If "TRUE" (default) data have been previously log transformed.

logbase
Numerical. Base for log transformation.

paired
Number of random pairs generated in the function, if set to NA (default), the odd integer closer to the square of the number of replicates is used.

cutoff
The pfp/pvalue threshold value used to select features.

method
If cutoff is provided, the method needs to be selected to identify metabolites. "pfp" uses percentage of false prediction, which is a default setting. "pval" uses p-values which is less stringent than pfp.

Value

A list with all results for Rank Product analysis including tables and plots.

References

Breitling, R., Armengaud, P., Amtmann, A., and Herzyk, P.(2004) Rank Products: A simple, yet powerful, new method to detect differentially regulated genes in replicated microarray experiments, FEBS Letter, 57383-92