Volcano Plot

PomaVolcano() generates a volcano plot from the PomaUnivariate(method = "ttest") result. The data can't have negative values.

PomaVolcano(
  data,
  pval = "raw",
  pval_cutoff = 0.05,
  adjust = "fdr",
  log2FC = 0.6,
  xlim = 2,
  labels = FALSE,
  paired = FALSE,
  var_equal = FALSE,
  interactive = FALSE,
  plot_title = TRUE
)

Arguments

data
A MSnSet object. First pData column must be the subject group/type. Only for two group data!

pval
Select a pvalue type to generate the volcano plot. Options are: "raw" and "adjusted".

pval_cutoff
Numeric. Define the pvalue cutoff (horizontal line).

adjust
Multiple comparisons correction method for t test result. Options are: "fdr", "holm", "hochberg", "hommel", "bonferroni", "BH" and "BY".

log2FC
Numeric. Define the log2 fold change cutoff (vertical lines).

xlim
Numeric. Define the limits for x axis.

labels
Logical that indicates if selected labels will be plotted or not. Defaul is FALSE.

paired
Logical that indicates if the data is paired or not.

var_equal
Logical that indicates if the data variance is equal or not.

interactive
Logical that indicates if an interactive plot will be plotted or not. Defaul is FALSE.

plot_title
Logical that indicates if title will be plotted or not. Defaul is TRUE.

Value

A ggplot2 object.