Collettion of Imputation Methods for Mass Spectrometry Data

PomaImpute() offers different methods to impute missing values in MS data.

PomaImpute(
  data,
  ZerosAsNA = FALSE,
  RemoveNA = TRUE,
  cutoff = 20,
  method = "knn"
)

Arguments

data
A MSnSet object. First pData column must be the subject group/type.

ZerosAsNA
Logical that indicates if the zeros in the data are missing values. Default is FALSE.

RemoveNA
Logical that indicates if those features with more than selected cutoff missing values in each group have to be removed. Default is TRUE.

cutoff
Numeric that indicates the percentage of missing values allowed in each group. If one of the groups have less missing values than selected cutoff value, these feature will not be removed.

method
Imputation method. Options are: "none", "half_min", "median", "mean", "min", "knn" and "rf". If "none", all missing values will be replaced by zero.

Value

A MSnSet object with cleaned data.

References

Armitage, E. G., Godzien, J., Alonso‐Herranz, V., López‐Gonzálvez, Á., & Barbas, C. (2015). Missing value imputation strategies for metabolomics data. Electrophoresis, 36(24), 3050-3060.

Examples

data("st000336")

PomaImpute(st000336, method = "knn")

#> MSnSet (storageMode: lockedEnvironment)
#> assayData: 30 features, 57 samples 
#>   element names: exprs 
#> protocolData: none
#> phenoData
#>   sampleNames: DMD004.1.U02 DMD005.1.U02 ... DMD173.1.U02 (57 total)
#>   varLabels: Group steroids
#>   varMetadata: labelDescription
#> featureData: none
#> experimentData: use 'experimentData(object)'
#> Annotation:  
#> - - - Processing information - - -
#> Imputed (knn): Sun Jun 21 20:55:28 2020 
#>  MSnbase version: 2.12.0