Univariate Statistical Methods for Mass Spectrometry Data

PomaUnivariate() allows users to perform different univariate statistical analysis on MS data.

PomaUnivariate(
  data,
  covariates = FALSE,
  method = "ttest",
  paired = FALSE,
  var_equal = FALSE,
  adjust = "fdr"
)

Arguments

data
A MSnSet object. First pData column must be the subject group/type.

covariates
Logical. If it's set to TRUE all metadata variables stored in pData will be used as covariables. Default = FALSE.

method
Univariate statistical method. Options are: "ttest", "anova", "mann" and "kruskal".

paired
Logical that indicates if the data is paired or not.

var_equal
Logical that indicates if the data variance is equal or not.

adjust
Multiple comparisons correction method. Options are: "fdr", "holm", "hochberg", "hommel", "bonferroni", "BH" and "BY".

Value

A data frame with results.

Author

Pol Castellano-Escuder

Examples

data("st000336")
data("st000284")

# ttest
st000336 %>% 
  PomaImpute() %>%
  PomaNorm() %>%
  PomaOutliers() %>%
  PomaUnivariate(method = "ttest")

#> method argument is empty! KNN will be used

#> method argument is empty! log_pareto will be used

#>                    mean_Controls   mean_DMD Fold_Change_Ratio
#> x1_methylhistidine   -0.40181818 0.16112821            -0.401
#> x3_methylhistidine   -0.41954545 0.19305128            -0.460
#> alanine              -0.31654545 0.10946154            -0.346
#> arginine             -0.11481818 0.06405128            -0.558
#> asparagine           -0.34754545 0.12764103            -0.367
#> aspartic_acid        -0.25427273 0.08887179            -0.350
#> glutamic_acid        -0.46127273 0.12617949            -0.274
#> glutamine            -0.29654545 0.14141026            -0.477
#> glycine              -0.25900000 0.15258974            -0.589
#> histidine            -0.24790909 0.13492308            -0.544
#> isoleucine           -0.46818182 0.13766667            -0.294
#> leucine              -0.46481818 0.15069231            -0.324
#> lysine               -0.28363636 0.11130769            -0.392
#> methionine           -0.41072727 0.14105128            -0.343
#> ornithine            -0.45990909 0.17287179            -0.376
#> phenylalanine        -0.43072727 0.14666667            -0.341
#> proline              -0.27363636 0.08666667            -0.317
#> serine               -0.37318182 0.16271795            -0.436
#> threonine            -0.38636364 0.16751282            -0.434
#> tryptophan           -0.59581818 0.17910256            -0.301
#> tyrosine             -0.41754545 0.14953846            -0.358
#> valine               -0.53409091 0.16689744            -0.312
#> x3_hba               -0.38536364 0.13910256            -0.361
#> a_kg                 -0.46000000 0.18256410            -0.397
#> citrate              -0.05527273 0.10261538            -1.857
#> fumarate             -0.31781818 0.05461538            -0.172
#> lactate              -0.59418182 0.19117949            -0.322
#> malate               -0.38072727 0.10428205            -0.274
#> pyruvate             -0.47500000 0.14946154            -0.315
#> succinate            -0.30318182 0.08666667            -0.286
#>                    Difference_Of_Means       pvalue    pvalueAdj
#> x1_methylhistidine               0.563 9.302122e-08 3.100707e-07
#> x3_methylhistidine               0.613 7.822367e-03 9.025808e-03
#> alanine                          0.426 6.617797e-04 8.631909e-04
#> arginine                         0.179 4.796275e-01 4.796275e-01
#> asparagine                       0.475 1.279748e-05 2.399528e-05
#> aspartic_acid                    0.343 3.302402e-02 3.538287e-02
#> glutamic_acid                    0.587 1.151414e-04 1.818022e-04
#> glutamine                        0.438 5.716908e-04 7.795784e-04
#> glycine                          0.412 3.978355e-03 4.972944e-03
#> histidine                        0.383 4.126907e-04 5.895582e-04
#> isoleucine                       0.606 1.947937e-08 1.369631e-07
#> leucine                          0.616 1.463078e-07 4.389233e-07
#> lysine                           0.395 2.922311e-02 3.247012e-02
#> methionine                       0.552 7.418771e-08 3.100707e-07
#> ornithine                        0.633 6.863241e-11 1.029486e-09
#> phenylalanine                    0.577 3.093799e-08 1.546899e-07
#> proline                          0.360 5.858996e-06 1.255499e-05
#> serine                           0.536 2.282719e-08 1.369631e-07
#> threonine                        0.554 3.918887e-06 9.043585e-06
#> tryptophan                       0.775 8.406570e-14 2.521971e-12
#> tyrosine                         0.567 3.013134e-06 7.532835e-06
#> valine                           0.701 2.359345e-10 2.359345e-09
#> x3_hba                           0.524 1.161299e-05 2.322598e-05
#> a_kg                             0.643 1.834223e-04 2.751334e-04
#> citrate                          0.158 1.707527e-01 1.766407e-01
#> fumarate                         0.372 2.190298e-05 3.865233e-05
#> lactate                          0.785 2.990736e-07 8.156552e-07
#> malate                           0.485 5.579962e-05 9.299937e-05
#> pyruvate                         0.624 8.744639e-08 3.100707e-07
#> succinate                        0.390 6.246847e-03 7.496217e-03

# ANOVA
st000284 %>% 
  PomaImpute() %>%
  PomaNorm() %>%
  PomaOutliers() %>%
  PomaUnivariate(method = "anova")

#> method argument is empty! KNN will be used

#> No missing values detected in your data

#> method argument is empty! log_pareto will be used

#>                                     mean_CRC  mean_Healthy   mean_Polyp
#> x1_methyladenosine              0.0584827586 -0.0612820513 -0.042708333
#> x1_methylhistamine             -0.0399310345  0.0127564103  0.071416667
#> x2_aminoadipate                -0.1451896552  0.0047435897 -0.028472222
#> x2_deoxyuridine                -0.1295344828  0.0338717949  0.055972222
#> x3_nitro_tyrosine              -0.0216034483 -0.0302435897  0.054736111
#> x4_pyridoxic_acid               0.0195517241 -0.0028461538 -0.016513889
#> x5_hydroxytryptophan            0.0267586207 -0.0028974359 -0.010486111
#> acetoacetate                    0.0844482759  0.0352435897  0.032611111
#> acetylcholine                  -0.0912413793 -0.0123974359  0.042277778
#> aconitate                      -0.0756206897 -0.0318461538 -0.010166667
#> adenosine                      -0.1715172414 -0.0356538462  0.180125000
#> adenylosuccinate                0.1693103448 -0.0397435897 -0.046652778
#> adipic_acid                     0.0631206897 -0.0631794872 -0.020833333
#> alanine                        -0.1000862069  0.0205384615  0.069236111
#> allantoin                       0.1663275862 -0.1144615385 -0.062736111
#> alpha_ketoglutaric_acid        -0.1102586207  0.0350512821  0.039069444
#> aminoisobutyrate               -0.0107241379 -0.0066153846 -0.030625000
#> amp                             0.0008793103 -0.0054487179  0.004333333
#> anthranilate                   -0.0501034483 -0.0276666667  0.038583333
#> arginine                        0.0621724138 -0.0358205128  0.002236111
#> asparagine                     -0.0476551724  0.0107435897  0.033291667
#> aspartic_acid                   0.1574655172 -0.0645897436 -0.088097222
#> betaine                        -0.0306206897 -0.0043717949  0.016388889
#> biotin                         -0.0925172414 -0.0129615385  0.104694444
#> carnitine                      -0.0579655172 -0.0144230769  0.077236111
#> choline                        -0.0344310345 -0.0316282051 -0.005541667
#> citraconic_acid                -0.0971379310 -0.0343717949  0.010208333
#> citrulline                     -0.0054310345 -0.0395769231  0.011694444
#> creatine                        0.0570862069 -0.0440128205  0.074180556
#> creatinine                     -0.1063965517  0.0023461538  0.004416667
#> cystamine                       0.0395517241 -0.0561025641 -0.062291667
#> cystathionine                   0.1061896552 -0.1569358974 -0.054597222
#> cystine                         0.0278103448 -0.0392051282  0.009833333
#> cytidine                        0.0646379310 -0.0519615385 -0.043763889
#> cytosine                        0.0302586207  0.0573333333 -0.157888889
#> d_ga3p_dhap                     0.0043448276 -0.0112820513  0.002097222
#> dimethylglycine                -0.1992413793  0.0703589744  0.136638889
#> d_leucic_acid                   0.1627413793 -0.0264871795 -0.108236111
#> dtmp                           -0.0169655172  0.0077179487  0.010208333
#> epinephrine                     0.0525517241 -0.0362948718 -0.019277778
#> erythrose                       0.0661206897 -0.0550769231 -0.006319444
#> f16bp_f26bp                     0.0032758621 -0.0287820513  0.030250000
#> fructose                        0.0591551724 -0.0336794872  0.004680556
#> fumaric                         0.0639137931 -0.0476153846 -0.010041667
#> g16bp                           0.0401896552 -0.0554743590  0.033208333
#> gama_aminobutyrate             -0.1448965517  0.0247692308  0.066805556
#> glucoronate                    -0.1147758621 -0.0344743590 -0.042250000
#> glucose                         0.0567241379 -0.0566282051  0.034013889
#> glutamic_acid                   0.1344482759 -0.0539230769 -0.051180556
#> glutamine                      -0.1208448276  0.0393461538  0.047625000
#> glutaric_acid                   0.0591206897 -0.0236153846 -0.095333333
#> glyceraldehyde                  0.2206724138 -0.0700384615 -0.095902778
#> glycerate                      -0.0835862069  0.0028076923  0.042527778
#> glycerol_3_p                   -0.0745517241  0.0100256410  0.055513889
#> glycine                        -0.0591551724 -0.0204743590  0.004736111
#> glycochenodeoxycholate          0.3128103448 -0.1990769231 -0.182472222
#> glycocholate                    0.3103103448 -0.1907051282 -0.207333333
#> guanidinoacetate               -0.0336206897  0.0159871795  0.040138889
#> guanosine                       0.1336379310  0.1322179487  0.054222222
#> histidine                      -0.1661551724  0.0984102564  0.052333333
#> homogentisate                  -0.0858965517  0.0389230769  0.040500000
#> homovanilate                    0.0021551724 -0.0415641026  0.035166667
#> hydroxyproline_aminolevulinate  0.1209310345 -0.0899743590 -0.085569444
#> hypoxanthine                   -0.0609310345  0.0181410256  0.066541667
#> hyppuric_acid                   0.2669310345 -0.1558205128 -0.216958333
#> imp                            -0.0576206897  0.1274102564 -0.077500000
#> inosine                         0.0292931034  0.0381410256  0.104319444
#> kynorenate                      0.0933793103 -0.0814615385 -0.074111111
#> lactate                         0.0756206897 -0.0635256410  0.024680556
#> leucine_iso_leucine            -0.0788275862  0.0108589744  0.075527778
#> linoleic_acid                  -0.1872241379  0.0193461538  0.107402778
#> linolenic_acid                 -0.2517931034  0.0685384615  0.117375000
#> l_kynurenine                    0.1049827586 -0.0526666667 -0.048416667
#> lysine                         -0.1534655172  0.0516025641  0.106069444
#> malate                         -0.0769827586 -0.0346410256  0.009083333
#> maleic_acid                     0.1033448276 -0.0625897436 -0.021263889
#> malonic_acid_3hba              -0.2631206897  0.1192692308  0.066194444
#> margaric_acid                  -0.1032931034  0.0091923077  0.048583333
#> methionine                     -0.1284310345  0.0492692308  0.096513889
#> methyl_succinate               -0.0682931034  0.0235128205  0.034250000
#> n2_n2_dimethylguanosine         0.0316551724 -0.0712820513 -0.067736111
#> n_acetyl_glycine               -0.2122241379  0.0436666667  0.062375000
#> niacinamide                    -0.0608965517 -0.0358076923  0.042347222
#> oh_phenylpyruvate               0.0258448276 -0.0453846154  0.027527778
#> ornithine                       0.0084137931 -0.0548846154  0.025222222
#> orotate                         0.0367068966 -0.0383076923 -0.124736111
#> oxalic_acid                     0.0855862069 -0.0795897436  0.036875000
#> oxaloacetate                    0.0153620690  0.0078717949 -0.107388889
#> pentothenate                   -0.1033275862 -0.0097179487  0.048805556
#> pep                            -0.1194137931  0.0119230769  0.068041667
#> phenylalanine                  -0.0032413793 -0.0025384615  0.018708333
#> proline                         0.0732931034 -0.0653974359 -0.015680556
#> propionate                     -0.0250862069  0.0115128205  0.021916667
#> pyridoxal_5_p                   0.0141896552  0.0500897436  0.104305556
#> pyroglutamic_acid              -0.0025000000 -0.0190769231 -0.027013889
#> pyruvate                        0.1110172414 -0.0996282051  0.034680556
#> reduced_glutatione             -0.0346206897  0.0481025641 -0.013916667
#> salicylurate                    0.0439827586 -0.0295641026  0.020458333
#> serine                         -0.0346551724  0.0005128205  0.027513889
#> shikimic_acid                   0.0552586207  0.0018205128 -0.023416667
#> sorbitol                       -0.0256206897  0.0077179487  0.065569444
#> succinate_methylmalonate       -0.0222068966 -0.0448205128 -0.011666667
#> taurine                        -0.0306724138  0.0029230769  0.079694444
#> threonine                      -0.0541724138  0.0203333333  0.072708333
#> trimethylamine_n_oxide          0.0630517241 -0.0555384615 -0.144722222
#> tryptophan                     -0.0196896552  0.0202435897  0.090111111
#> tyrosine                       -0.0362586207  0.0201410256  0.056388889
#> urate                          -0.0853275862  0.0323846154  0.040208333
#> uridine                        -0.1451034483  0.0187948718  0.111666667
#> valine                         -0.0526206897  0.0336282051  0.047361111
#> xanthine                       -0.1158448276 -0.0107692308  0.087138889
#> xanthosine                     -0.0441551724 -0.0589615385 -0.050069444
#> xanthurenate                   -0.0680689655  0.0261666667  0.094069444
#>                                      pvalue    pvalueAdj
#> x1_methyladenosine             4.536408e-02 1.114378e-01
#> x1_methylhistamine             1.580891e-01 2.666284e-01
#> x2_aminoadipate                5.326932e-02 1.228456e-01
#> x2_deoxyuridine                4.109506e-05 5.031429e-04
#> x3_nitro_tyrosine              1.580894e-01 2.666284e-01
#> x4_pyridoxic_acid              4.832485e-01 6.205350e-01
#> x5_hydroxytryptophan           5.423561e-01 6.733882e-01
#> acetoacetate                   8.616282e-01 8.932476e-01
#> acetylcholine                  1.170646e-01 2.236269e-01
#> aconitate                      5.875096e-01 6.942735e-01
#> adenosine                      1.888282e-02 5.766916e-02
#> adenylosuccinate               1.361886e-02 4.526270e-02
#> adipic_acid                    4.479071e-01 5.954789e-01
#> alanine                        7.458530e-03 3.159687e-02
#> allantoin                      1.797733e-03 1.269649e-02
#> alpha_ketoglutaric_acid        5.294223e-04 4.273194e-03
#> aminoisobutyrate               8.487680e-01 8.932476e-01
#> amp                            9.532497e-01 9.617609e-01
#> anthranilate                   5.898253e-01 6.942735e-01
#> arginine                       1.249544e-01 2.241246e-01
#> asparagine                     2.252417e-01 3.439502e-01
#> aspartic_acid                  4.882716e-03 2.298946e-02
#> betaine                        6.529243e-01 7.378044e-01
#> biotin                         1.098738e-03 8.277158e-03
#> carnitine                      1.202954e-02 4.247933e-02
#> choline                        8.537685e-01 8.932476e-01
#> citraconic_acid                2.380732e-01 3.586969e-01
#> citrulline                     6.352719e-01 7.325074e-01
#> creatine                       2.105085e-01 3.338411e-01
#> creatinine                     2.034551e-02 6.050111e-02
#> cystamine                      4.392001e-01 5.954789e-01
#> cystathionine                  4.148625e-03 2.038238e-02
#> cystine                        2.176889e-01 3.369705e-01
#> cytidine                       2.682480e-01 3.936626e-01
#> cytosine                       5.687629e-01 6.837256e-01
#> d_ga3p_dhap                    8.416138e-01 8.932476e-01
#> dimethylglycine                7.106257e-06 1.527649e-04
#> d_leucic_acid                  2.279582e-02 6.282749e-02
#> dtmp                           6.421840e-01 7.329979e-01
#> epinephrine                    1.234946e-01 2.241246e-01
#> erythrose                      9.751357e-02 2.020045e-01
#> f16bp_f26bp                    4.479266e-01 5.954789e-01
#> fructose                       6.105609e-01 7.112720e-01
#> fumaric                        9.742969e-02 2.020045e-01
#> g16bp                          4.721002e-02 1.135049e-01
#> gama_aminobutyrate             8.987009e-05 8.462767e-04
#> glucoronate                    5.624872e-01 6.834522e-01
#> glucose                        5.117904e-02 1.204840e-01
#> glutamic_acid                  2.572918e-02 6.922375e-02
#> glutamine                      4.897852e-05 5.031429e-04
#> glutaric_acid                  6.213796e-02 1.404318e-01
#> glyceraldehyde                 4.661308e-07 3.103475e-05
#> glycerate                      2.756094e-01 3.992804e-01
#> glycerol_3_p                   1.185341e-01 2.236269e-01
#> glycine                        5.482453e-01 6.733882e-01
#> glycochenodeoxycholate         4.568408e-05 5.031429e-04
#> glycocholate                   3.079631e-03 1.808852e-02
#> guanidinoacetate               4.305532e-01 5.933233e-01
#> guanosine                      8.044139e-01 8.740267e-01
#> histidine                      5.183795e-06 1.464422e-04
#> homogentisate                  1.581976e-02 5.101992e-02
#> homovanilate                   3.161451e-01 4.466178e-01
#> hydroxyproline_aminolevulinate 3.956460e-03 2.032182e-02
#> hypoxanthine                   1.754784e-01 2.916039e-01
#> hyppuric_acid                  8.111413e-06 1.527649e-04
#> imp                            9.527426e-03 3.844997e-02
#> inosine                        7.751630e-01 8.622998e-01
#> kynorenate                     7.549696e-03 3.159687e-02
#> lactate                        3.870405e-02 9.820174e-02
#> leucine_iso_leucine            1.325641e-02 4.526270e-02
#> linoleic_acid                  2.515888e-04 2.186887e-03
#> linolenic_acid                 6.592401e-07 3.103475e-05
#> l_kynurenine                   2.151269e-02 6.077334e-02
#> lysine                         8.239314e-07 3.103475e-05
#> malate                         3.853944e-01 5.376490e-01
#> maleic_acid                    5.767748e-03 2.607022e-02
#> malonic_acid_3hba              2.124229e-03 1.411988e-02
#> margaric_acid                  1.625414e-02 5.101992e-02
#> methionine                     1.878475e-05 3.032396e-04
#> methyl_succinate               7.664566e-02 1.665569e-01
#> n2_n2_dimethylguanosine        1.099953e-01 2.219548e-01
#> n_acetyl_glycine               2.462168e-03 1.545694e-02
#> niacinamide                    4.584010e-01 5.982889e-01
#> oh_phenylpyruvate              2.506194e-01 3.726315e-01
#> ornithine                      3.161896e-01 4.466178e-01
#> orotate                        3.645455e-02 9.579916e-02
#> oxalic_acid                    1.037876e-02 3.928452e-02
#> oxaloacetate                   1.288436e-01 2.274895e-01
#> pentothenate                   2.085314e-01 3.338411e-01
#> pep                            3.361583e-03 1.808852e-02
#> phenylalanine                  8.802792e-01 9.042868e-01
#> proline                        3.910689e-02 9.820174e-02
#> propionate                     7.963528e-01 8.736686e-01
#> pyridoxal_5_p                  5.119790e-01 6.428180e-01
#> pyroglutamic_acid              9.137684e-01 9.302327e-01
#> pyruvate                       2.142552e-02 6.077334e-02
#> reduced_glutatione             9.832080e-02 2.020045e-01
#> salicylurate                   5.079468e-01 6.428180e-01
#> serine                         4.606295e-01 5.982889e-01
#> shikimic_acid                  8.162879e-01 8.784813e-01
#> sorbitol                       2.127129e-01 3.338411e-01
#> succinate_methylmalonate       7.783591e-01 8.622998e-01
#> taurine                        1.130267e-01 2.236269e-01
#> threonine                      7.581686e-02 1.665569e-01
#> trimethylamine_n_oxide         1.195902e-01 2.236269e-01
#> tryptophan                     1.207189e-01 2.236269e-01
#> tyrosine                       2.091160e-01 3.338411e-01
#> urate                          1.077717e-02 3.928452e-02
#> uridine                        4.011986e-05 5.031429e-04
#> valine                         1.517422e-01 2.637980e-01
#> xanthine                       3.242901e-03 1.808852e-02
#> xanthosine                     9.671710e-01 9.671710e-01
#> xanthurenate                   1.056522e-02 3.928452e-02